kernel: moving lands to arch specific module
Splitting of memory is choosed based on MMU, and this is arch-specific.

Closes #179
