+++ kernel/src/paging/lands.rs
@@ -1,6 +1,8 @@
 //! Module describing the split between the UserSpace and KernelSpace,
 //! and a few functions to work with it.
 
+pub use super::arch::{KernelLand, UserLand, RecursiveTablesLand};
+
 use crate::mem::VirtualAddress;
 use super::PAGE_SIZE;
 use super::arch::ENTRY_COUNT;
@@ -73,49 +75,3 @@ pub trait VirtualSpaceLand {
         }
     }
 }
-
-/// The virtual memory belonging to kernel
-#[derive(Debug)] pub struct KernelLand;
-/// The virtual memory belonging to user
-#[derive(Debug)] pub struct UserLand;
-/// The virtual memory pointing to active page tables by recursion
-#[derive(Debug)] pub struct RecursiveTablesLand;
-
-// TODO: move KernelLand Userland RTL to arch-specific paging
-// BODY: They are arch dependant, we should stop trying defining them in an agnostic way,
-// BODY: even if they are expected to be the same for 32 bits architectures.
-// BODY:
-// BODY: Especially for the Recursive Tables Land. Even if it's a agnostic concept, its size in
-// BODY: virtual memory is mmu-dependant, and defined by the number of levels the mmu uses.
-
-// if 32 bit, we define UserLand and KernelLand here
-#[cfg(any(target_pointer_width = "32", test))]
-impl VirtualSpaceLand for UserLand {
-    const START: VirtualAddress = VirtualAddress(0x00000000);
-    const END:   VirtualAddress = VirtualAddress(0xbfffffff);
-}
-
-#[cfg(any(target_pointer_width = "32", test))]
-impl VirtualSpaceLand for KernelLand {
-    const START: VirtualAddress = VirtualAddress(0xc0000000);
-    const   END: VirtualAddress = VirtualAddress(0xffbfffff);
-}
-
-#[cfg(any(target_pointer_width = "32", test))]
-impl VirtualSpaceLand for RecursiveTablesLand {
-    const START: VirtualAddress = VirtualAddress(0xffc00000);
-    const   END: VirtualAddress = VirtualAddress(0xffffffff);
-}
-// else we do it in arch-specific implementations
-
-/// Assertions to check that Kernel/User pages falls on distinct page tables
-/// and also that they do not overlap.
-const_assert!(KernelLand::START.0 < KernelLand::END.0);
-const_assert!(UserLand::START.0 < UserLand::END.0);
-const_assert!(RecursiveTablesLand::START.0 < RecursiveTablesLand::END.0);
-// TODO: Const FN sucks! Check that the kernelland and userland don't overlap.
-//const_assert!(::core::cmp::max(KernelLand::start_addr(), UserLand::start_addr()) >=
-//              ::core::cmp::min(KernelLand::end_addr(),   UserLand::end_addr()));
-
-const_assert!(KernelLand::START.0 % (ENTRY_COUNT * PAGE_SIZE) == 0);
-const_assert!(UserLand::START.0   % (ENTRY_COUNT * PAGE_SIZE) == 0);
